Revert "Refactor - extract Vs dependent code"
This reverts commit d306f730730f01a5b56bebd908ddc8a445f0685d.
